requirements.txt
numpy==1.22.4
opencv_python==4.5.4.60
Pillow==9.1.1
scipy==1.8.1
torch==1.8.0
torchvision==0.9.0
tqdm==4.64.0
